/guides/select is DOOOONE!
Hardest by far, when it comes to the HTML, CSS, and also some JavaScript. But it turned out pretty well so it wasn't all for nothing. We'll start rolling out some new pages later on. Have  a good night! We'll be going through most of the night on this, this site is a full-time project at this point.